+++ server/src/routes/admin.js
@@ -29,6 +29,7 @@ import {
 } from '../services/offerLetter.js';
 import { previewCycleEventCopy, commitCycleEventCopy } from '../services/eventCopy.js';
 import { createCalendarEvent, updateCalendarEvent, cancelCalendarEvent } from '../services/google/calendar.js';
+import { setEventInvitees, syncEventCalendarInvite } from '../services/google/eventCalendarSync.js';
 
 const router = express.Router();
 
@@ -1466,71 +1467,6 @@ router.get('/profile', async (req, res) => {
 });
 
 // Event Management Routes
-
-// Resolves who should receive the Google Calendar invite for an event: every ADMIN (always)
-// plus whichever MEMBERs were manually selected as invitees for this specific event.
-async function getEventAttendeeEmails(eventId) {
-  const [admins, invitees] = await Promise.all([
-    prisma.user.findMany({ where: { role: 'ADMIN' }, select: { email: true } }),
-    prisma.eventInvitee.findMany({ where: { eventId }, select: { user: { select: { email: true } } } })
-  ]);
-  const emails = new Set([...admins.map((a) => a.email), ...invitees.map((i) => i.user.email)]);
-  return [...emails];
-}
-
-// Filters a list of candidate invitee IDs down to real MEMBER-role users, so candidates/admins
-// can't be added to the invitee join table (admins are already always invited).
-async function resolveMemberInviteeIds(memberInviteeIds) {
-  if (!Array.isArray(memberInviteeIds) || memberInviteeIds.length === 0) return [];
-  const members = await prisma.user.findMany({
-    where: { id: { in: memberInviteeIds }, role: 'MEMBER' },
-    select: { id: true }
-  });
-  return members.map((m) => m.id);
-}
-
-async function setEventInvitees(eventId, memberInviteeIds) {
-  const validIds = await resolveMemberInviteeIds(memberInviteeIds);
-  await prisma.$transaction([
-    prisma.eventInvitee.deleteMany({ where: { eventId } }),
-    ...(validIds.length
-      ? [prisma.eventInvitee.createMany({ data: validIds.map((userId) => ({ eventId, userId })) })]
-      : [])
-  ]);
-  return validIds;
-}
-
-// Creates or updates the Google Calendar invite for an event. Never throws — a calendar/API
-// failure shouldn't block event create/update; callers get back { googleCalendarEventId, calendarError }.
-async function syncEventCalendarInvite(event) {
-  try {
-    const [attendeeEmails, cycle] = await Promise.all([
-      getEventAttendeeEmails(event.id),
-      prisma.recruitingCycle.findUnique({ where: { id: event.cycleId }, select: { name: true } })
-    ]);
-
-    const eventDetails = {
-      eventName: event.eventName,
-      eventLocation: event.eventLocation,
-      eventStartDate: event.eventStartDate,
-      eventEndDate: event.eventEndDate,
-      cycleName: cycle?.name
-    };
-
-    const googleCalendarEventId = event.googleCalendarEventId
-      ? await updateCalendarEvent(event.googleCalendarEventId, eventDetails, attendeeEmails)
-      : await createCalendarEvent(eventDetails, attendeeEmails);
-
-    if (googleCalendarEventId && googleCalendarEventId !== event.googleCalendarEventId) {
-      await prisma.events.update({ where: { id: event.id }, data: { googleCalendarEventId } });
-    }
-
-    return { googleCalendarEventId, attendeeCount: attendeeEmails.length };
-  } catch (error) {
-    console.error(`[Calendar] Failed to sync invite for event ${event.id}:`, error);
-    return { googleCalendarEventId: event.googleCalendarEventId || null, calendarError: error.message };
-  }
-}
 
 // Get all events
 router.get('/events', async (req, res) => {
